Notice when users look at dashlet and there are replies they can't delete

+++ sugarcrm/modules/SugarFeed/SugarFeed.php
@@ -351,7 +351,7 @@ class SugarFeed extends Basic {
 			$data['NAME'] = '';
 			return $data;
 		}
-		if(is_admin($GLOBALS['current_user']) || $data['CREATED_BY'] == $GLOBALS['current_user']->id) {
+		if(is_admin($GLOBALS['current_user']) || (isset($data['CREATED_BY']) && $data['CREATED_BY'] == $GLOBALS['current_user']->id) ) {
             $delete = ' - <a id="sugarFeedDeleteLink'.$data['ID'].'" href="#" onclick=\'SugarFeed.deleteFeed("'. $data['ID'] . '", "{this.id}"); return false;\'>'. $GLOBALS['app_strings']['LBL_DELETE_BUTTON_LABEL'].'</a>';
         }
 		$data['NAME'] .= $data['DESCRIPTION'];
